git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2701 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/combigrid/utils/CombigridLevelVector.cpp
@@ -142,7 +142,7 @@ CombigridLevelVector CombigridLevelVector::getCombiLevels(CombigridLevelVector i
 
 CombigridLevelVector CombigridLevelVector::getChanges(std::vector<int> in){
 	CombigridLevelVector inVector(in);
-	CombigridLevelVector unity(in.size());
+	CombigridLevelVector unity( static_cast<int>( in.size() ) );
 //	CombigridLevelVector current(levelVec_,coef_);
 	std::vector<CombigridLevelVector> currentVec=(*this).split();
 	for(unsigned int i=0;i<levelVec_.size();i++){
